refactor[#8] : FSD 구조에서 views 폴더 지우고 nextjs 의 앱라우팅 기반 따르도록 변경
- 추후 다른 nextjs 최적화 할 때 기능이 복잡해지는 것을 염려
