--- blog/src/views/main/index.ts
@@ -1,2 +1,0 @@
-export * from "./layout";
-export * from "./page";
