+++ blog/src/widgets/navigate/lib/index.ts
